Move token endpoint to auth from users
