+++ local-requirements.txt
@@ -1,3 +1,4 @@
 Django>=4.0.6,<4.1
 djangorestframework>=3.13.1,<3.14
+drf-spectacular>=0.23.1,<0.24
 flake8>=5.0.4,<5.1
