+++ ther00m/settings.py
@@ -32,7 +32,8 @@ INSTALLED_APPS = [
     'django.contrib.admin', 'django.contrib.auth',
     'django.contrib.contenttypes', 'django.contrib.sessions',
     'django.contrib.messages', 'django.contrib.staticfiles', 'users',
-    'rest_framework', 'drf_spectacular', 'rest_framework.authtoken'
+    'rest_framework', 'drf_spectacular', 'rest_framework.authtoken',
+    'user_auth'
 ]
 
 MIDDLEWARE = [
