+++ ther00m/urls.py
@@ -24,5 +24,6 @@ urlpatterns = [
          SpectacularSwaggerView.as_view(url_name='api-schema'),
          name='api-docs'),
     path('api/users/', include('users.urls')),
-    path('api/health/', include('health.urls'))
+    path('api/health/', include('health.urls')),
+    path('api/auth/', include('user_auth.urls'))
 ]
