+++ users/urls.py
@@ -5,6 +5,5 @@ app_name = "users"
 
 urlpatterns = [
     path('create/', views.CreateUserView.as_view(), name='create'),
-    path('token/', views.CreateTokenView.as_view(), name='token'),
     path('me/', views.ManageUserView.as_view(), name='me')
 ]
