+++ users/views.py
@@ -1,8 +1,6 @@
 from rest_framework import generics, authentication, permissions
-from rest_framework.authtoken.views import ObtainAuthToken
-from rest_framework.settings import api_settings
 
-from users.serializers import UserSerializer, AuthTokenSerializer
+from users.serializers import UserSerializer
 
 
 class CreateUserView(generics.CreateAPIView):
@@ -21,7 +19,7 @@ class ManageUserView(generics.RetrieveUpdateAPIView):
         return self.request.user
 
 
-class CreateTokenView(ObtainAuthToken):
-    """Create a new auth token for a user."""
-    serializer_class = AuthTokenSerializer
-    renderer_classes = api_settings.DEFAULT_RENDERER_CLASSES
+# class CreateTokenView(ObtainAuthToken):
+#     """Create a new auth token for a user."""
+#     serializer_class = AuthTokenSerializer
+#     renderer_classes = api_settings.DEFAULT_RENDERER_CLASSES
